move to qmcpy/kernel_methods
