+++ qmcpy/__init__.py
@@ -2,8 +2,7 @@ from .discrete_distribution import *
 from .true_measure import *
 from .integrand import *
 from .stopping_criterion import *
-from .util import kernel_methods
-from .util.kernel_methods import (
+from .kernel_methods import (
     fftbr,ifftbr,fwht,omega_fftbr,omega_fwht,
     fftbr_torch,ifftbr_torch,fwht_torch,omega_fftbr_torch,omega_fwht_torch,
     kernel_shift_invar,kernel_dig_shift_invar,kernel_si,kernel_dsi)
