+++ qmcpy/kernel_methods/__init__.py
