+++ qmcpy/kernel_methods/dig_shift_invar_ops.py
