+++ qmcpy/kernel_methods/ft.py
