+++ qmcpy/kernel_methods/shift_invar_ops.py
