+++ qmcpy/stopping_criterion/cub_qmc_bayes_net_g.py
@@ -1,7 +1,8 @@
 from .abstract_cub_bayes_ld_g import AbstractCubBayesLDG
 from ..discrete_distribution import DigitalNetB2
 from ..integrand import Keister,BoxIntegral,Genz,SensitivityIndices
-from ..util import fwht,omega_fwht,MaxSamplesWarning, ParameterError, ParameterWarning, NotYetImplemented
+from ..kernel_methods import fwht,omega_fwht
+from ..util import MaxSamplesWarning, ParameterError, ParameterWarning, NotYetImplemented
 import ctypes
 import numpy as np
 from time import time
