+++ qmcpy/stopping_criterion/cub_qmc_lattice_g.py
@@ -4,7 +4,8 @@ from ..true_measure import Gaussian, Uniform
 from ..integrand import Keister, BoxIntegral, CustomFun
 from ..integrand.genz import Genz
 from ..integrand.sensitivity_indices import SensitivityIndices
-from ..util import fftbr,omega_fftbr,ParameterError
+from ..kernel_methods import fftbr,omega_fftbr
+from ..util import ParameterError
 import numpy as np
